requirements.txt
pandas
openpyxl
pyarrow

requests
beautifulsoup4
lxml

selenium
webdriver-manager

python-dotenv
openai